requirements.txt
asgiref==3.4.1
av==10.0.0
boto3==1.24.11
botocore==1.27.11
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
colorama==0.4.5
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0
django-heroku==0.3.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
easydict==1.10
filelock==3.4.2
Flask==2.2.2
ftfy==6.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.1
Pillow==8.4.0
platformdirs==2.4.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
regex==2023.8.8
requests==2.28.1
s3transfer==0.6.0
scipy==1.10.0
six==1.16.0
sqlparse==0.4.2
sympy==1.12
torch==2.0.1
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tqdm==4.66.1
typing_extensions==4.6.3
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.12.0
wcwidth==0.2.6
Werkzeug==2.2.2
whitenoise==6.0.0
